+++ samples/kotlin-inject/src/main/java/com/joshafeinberg/circuitkotlininject/sample/AppComponent.kt
@@ -1,0 +1,14 @@
+// Copyright (C) 2024 Slack Technologies, LLC
+// SPDX-License-Identifier: Apache-2.0
+package com.joshafeinberg.circuitkotlininject.sample
+
+import me.tatarka.inject.annotations.Component
+import me.tatarka.inject.annotations.Provides
+
+@Component
+abstract class AppComponent {
+  @Provides
+  fun providesString(): String {
+    return "Injected String!"
+  }
+}
