+++ samples/kotlin-inject/src/main/java/com/joshafeinberg/circuitkotlininject/sample/AppScope.kt
@@ -1,0 +1,7 @@
+// Copyright (C) 2024 Slack Technologies, LLC
+// SPDX-License-Identifier: Apache-2.0
+package com.joshafeinberg.circuitkotlininject.sample
+
+import me.tatarka.inject.annotations.Scope
+
+@Scope annotation class AppScope
